Bug 91124
Can't view inline image attachments in forwarded emails


What I noticed is that ZmComposeView.prototype._restoreMultipartRelatedImages already included code, for bug 85129  that happens to also deal with the case of this bug. However, there was a slight bug/typo there, fixed here - if the src is of the CSFE_MSG_FETCHER_URI format (e.g. localhost....) parseQueryString of src and not of dfsrc like it was doing - the thing is in many cases, dfsrc is the same (seems to be when dfsrc is given the value from attribute "data-mce-src" in the same method above. However data-mce-src is not always there and anyway so sometimes dfsrc is null and NOT the same as src. So I fixed that. Again, sometimes this case happens and sometimes it doesn't even with seemingly exactly the same repro steps. I tested as much as I could and it all seems to work now.

And again the original repro steps:
1. send a message from user1 to user2 that includes (a) an inline attachment and (b) an external image (via copy-paste from web for example).
2. in user2 view the message and click "display images" to show the external image (unless you are set to always display, in which case this bug probably also reproduces, but I haven't checked).
3. click "forward" and forward to user3
4. in user3, the inline image does not appear, and in fact it is not included as a part in the "show original" (as it was not included in step 3).

http://bugzilla.zimbra.com/show_bug.cgi?id=91124

Copied from Perforce
 Change: 514070