Bug 84779
new compose tab lost


I can't repro this bug, and this is probably not a solution to it. Since I can't repro I'be been trying to look at the code carefully for any possible vulnerability that could cause it. I found this issue but I don't think it causes any bug, but the code is a little fragile. (again I don't think it's broken currently, other things make it work, and I also don't think it could have anything to do with this tab being lost issue, but I'm really lost as to thinking of possible reasons).

This fix is to make sure:
1. this._dontSavePreHide is reset to false in _preHideCallback at the top so it's not possible this method exits without resetting it.
2. there's also no reason to try to save a draft (by calling _autoSaveCallback) in the case of sending the message (not saving a draft) - and again it was not saving, since it was not dirty, but it's safer to set _dontSavePreHide in this case too, so I added it. (when popping the view in _handleResponseSendMsg).

http://bugzilla.zimbra.com/show_bug.cgi?id=84779

Copied from Perforce
 Change: 514114