bug 90492 - less reset size calls
Currently, the ZmHtmlEditor::_resetSize() issues a delayed call on
each invocation. As we get at least one call each time the window is
resized, this can add up to quite a lot of reocurring calls;
especially if the window is too small to show the editor in full, in
which case we just continue postponing the calls.

To address this, stash the timeout ID on the editor instance, and
clear it whenever calling resetSize() anew, thus ensuring that we only
ever have a single call pending.

http://bugzilla.zimbra.com/show_bug.cgi?id=90492

Copied from Perforce
 Change: 514157