Bug 88228
Collapsing new message in conv view sometimes makes message disappear


One possible repro steps (easiest):
1. Mark a message or a conv as "unread" (while it's viewed in the reading pane - do not click on it afterwards).
2. Collapse the message (or one of the messages in the conv if marked the entire conv unread).

Another possible repro is if the message you collapse is in the trash.

And yet another possible scenario is if either the "mark read" setting in the preferences is to delay, or the server is very slow, and you collapse the message before the server returned from setting all the messages to "read" (once you view them, they are all read)

This is a regression from bug 82989. Bug 82989 tried to figure out if a child of the capsule view is the header, by using the class (comparing the entire string). The problem is the header class in those cases is "Conv2MsgHeader Unread" (or "Conv2MsgHeader Trash") so the comparison fails. A better solution would have been to do indexOf, which would have worked, but this is not that great, to use css class to figure out a type of child. So I changed that to use pure JS indication.

http://bugzilla.zimbra.com/show_bug.cgi?id=88228

Copied from Perforce
 Change: 514319