Bug: 86946 - Single Sign On Integration with Social Platform
Look for community URL (which includes OAuth signature) in GetInfoResponse.
If it's there, it'll be used to set the source URL for the Community tab iframe.
If not, we revert to old behavior and look at zimbraFeatureSocialExternalURL.

We no longer look at zimbraFeatureSocialExternalEnabled. The presence of the
URL means the feature is enabled.

http://bugzilla.zimbra.com/show_bug.cgi?id=86946

Copied from Perforce
 Change: 514355