Bug: 89902 - Script Error: this._evt is undefined
When offline mode is enabled and actions performed upon appts, the delta sync requests will load modified appts, which replace entries in the AppCtxt cache.  The appts are cleaned prior to being stored in the offline db, leaving a cached version that will be missing expected subobjects.

Simplest solution - suppress appCtxt caching for the offline appt reads upon sync.  When online, the normal server reads and caching work as expected, and when offline the appts are accessed from the db (caching for appts only being used for notifications, which doesn't apply offline).

http://bugzilla.zimbra.com/show_bug.cgi?id=89902

Copied from Perforce
 Change: 514642