Bug: 92176 - script error deleting mail item within a saved search
In ZmSearch.js the function ZmParsedQuery.getMatchFunction is taking the tokens:
0: arg: "unread",             op: "is",   type: "TERM"
1: arg: undefined,           op: "(",    type: "GROUP"
2: arg: "activity stream",  op: "in",  type: "TERM"
3: arg: undefined,           op: "or",  type: "COND"
4: arg: "inbox",               op: "in",  type: "TERM"
5: arg: undefined,           op: ")",    type: "GROUP"

And producing: 
  "return Boolean(item.isUnread(((item.type === ZmItem.CONV) ? item.folders && item.folders['624'] : item.folderId === '624') ||
   ((item.type ===  ZmItem.CONV) ? item.folders && item.folders['2'] : item.folderId === '2')))"
So that the boolean item.isUnread looks like a function invocation.

It looks like CL 495794 set the code that adds an 'AND' to compare using ===, which is too strict (comparing text and a ZmSearchToken).  Reverting it to '==' comparisons.

http://bugzilla.zimbra.com/show_bug.cgi?id=92176

Copied from Perforce
 Change: 514845