Bug 79656
The screen layout not fully displayed during editing Contact Group


Fixed the height sizing of the results contact list. Previously it had lots of hacky magic numbers based on all kinds of checks (all in ZmGroupView.prototype.setBounds) - I changed it to calculate the hight for it by reducing real heights of DOM elements from the total view height.

Made sure it works for edit DL too which uses the same code... slight difference there is the tabs, which take space, so took care of that too.

http://bugzilla.zimbra.com/show_bug.cgi?id=79656

Copied from Perforce
 Change: 515049