Bug 92724
Selecting contact group from shared address book in autocomplete can yield partial members


This is a simple solution to what's happening. The shared group is cached in a non-loaded (non dereferenced) state.  Not sure it should be cached that way, and it probably wasn't at some point, as the following comment in ZmContact says:

"// shared contacts are never fully loaded since we never cache them" not precise but interesting.

But at that case isLoaded is set to false, and so my solution is in the ZmAutocompleteMatch constractor, treat this case the same as if the group is not in the cache, resulting in what we want - a GetContactsRequest call to load and deref the members.

http://bugzilla.zimbra.com/show_bug.cgi?id=92724

Copied from Perforce
 Change: 515574