Bug 92950 - Save or Save & Close button is not working for new document in briefcase
Addressing an edge use case when a new document is open, nothing is typed in the body or title and "save" or "save & close" is pressed.

The test if(!_docModified && !_docNameModified) was going through since the new document didn't have any new content and was considrered not modified so the window was getting closed without any status message or warning. Added a check to see if there is a warning message to be displayed to the user before closing the window.

Now, if the new document is being created without a title, and is attempted to be saved, the dialog will prompt user "Document must have a file name."

http://bugzilla.zimbra.com/show_bug.cgi?id=92950

Copied from Perforce
 Change: 516450