Bug 92295
Modify Ajax Client to handle CSRF token


Improve my previous fix to this and also to the solution for the auth token expiration for child window.

Previous comment:

Fix issue that exists in some cases (so far I only see it in briefcase) in child windows. If the child window is initiating a request it too should use the csrfToken in the request, of course. I previously tested the mail new window (compose and view) cases and they worked since the child called the parent window to perform the requests, so I assumed all child windows call the parent for sending requests. But when looking at bug 92950, I noticed save stopped working due to this issue.

To make things more complicated, took me a while to figure out another problem, due to bug 92831 - which I re-opened and my fix here undoes what bug 92831 did, at least temporarily. (Not sure if bug 92831 is a real vulnerability).

http://bugzilla.zimbra.com/show_bug.cgi?id=92295

Copied from Perforce
 Change: 516458