bug 92861 - ease debugging TinyMCE in standard client
Normally, we embed the minified JavaScript sources for TinyMCE
directly in the JSP file, but they're somewhat hard to debug. Instead,
make ?dev=1 trigger TinyMCE's compiled inline mode -- the development
mode hardcodes the name of the development script and so doesn't
work. (We replace dots in filenames with underscores to make the
scripts work with AjxPackage.)

http://bugzilla.zimbra.com/show_bug.cgi?id=92861

Copied from Perforce
 Change: 516548