Bug 77635
refresh external imap account removes string query


Unless I'm missing something, the only minor bug is that the query disappears. This is for any folder, not just external, when clicking the "refresh" button.

Bug 56923 made a change to move the refresh button to the masthead, and when clicked, run the runRefresh method of all the active applications, so all are updated. This was done on purpose.

This causes the minor bug of losing the query, due to something that happens in the Calendar app - ZmCalViewController.prototype._refreshButtonListener, it clears the query in the search bar. This is since for Calendar, the "refresh" actually does clear the query, e.g. if it's a tag clicked on (tag:foo) or a search folder. "Refresh" of the calendar app (which again, is called for all the apps regardless of the current viewed app when the "refresh" was clicked) shows the default view again. (there was a bug recently discussing it, this is the only way to clear saved searches and tags).

This fix is to clear the search bar only if the calendar app is the current app. But we also need to clear the currentSearch and currentQuery of the app due to the weirdness of the Calendar app, that _showResults is not called for the full calendar case. (it's only called thus setting currentSearch and currentApp, in case of tag or saved search clicked) - THIS is actually a bigger bug, that this solves! I knew it was worth it to spend the time investigating this :) - the bug is the following repro steps:
1. go to calendar tab
2. assume you already have a tag in the tree, click on it. Notice the query in the search bar (assuming the pref is set to show advanced query) shows tag:foo
3. click the refresh button. Notice the view shows again ALL the items, not just the ones from the tag, as expected. Also the search bar is cleared, as expected.
4. click on mail tab
5. click on calendar tab again.
Result - the search bar is showing "tag:foo" again, which is wrong since the calendar shows everything (as expected).

http://bugzilla.zimbra.com/show_bug.cgi?id=77635

Copied from Perforce
 Change: 517092