Bug 93453
Viewing mail in search throwing errors and blank views


This solves the exception that happens after clicking on an "empty" conv (most likely a draft) and THEN clicking on another conv.

It's a regression from bug 77255.

Bug 77255 solution had several issues.
1. The relevant code was executed regardless if the same conv was clicked again (this is the case that it should be called).
2. Some defensive null checks were not there, though after fixing (1) I'm not sure it's really required but did it anyway.

In addition, I don't understand how it (expanding the first message) worked after the fix at all, with the repro steps described in bug 77255. It probably didn't, due to another bug I fixed here - when getting a new message visa a notification (say by clicking the "refresh" button as suggested), _msgViewList, which is supposed to be the msg IDs in the order they are displayed (the name is bad, but not fixing it now), loses the order as the ID of the new message is added at the end always (in ZmConvView2.prototype._renderMessage) - so solved this too:

3. in ZmConvView2.prototype._renderMessage, if an index param is passed, use it to set the correct location for the new message ID in _msgViewList.

http://bugzilla.zimbra.com/show_bug.cgi?id=93453

Copied from Perforce
 Change: 517549