Bug 93453
Viewing mail in search throwing errors and blank views


This is the blank view issue, which is more general than that.

Had a long discussion with Josh about it, as it's very tricky to get all the cases right.

We decided for now to not omit drafts anywhere. Drafts are part of the conversation, even though they were not sent yet, as at least server-wise we have to specify what conv they belong to when we save them.
If we decide we don't want them in regular folders such as Inbox, I can easily make it so it behaves like trash, which is the next thing I fixed and will discuss:
For trash, the case where a conv could end up empty in search results is only if the user checks the "include trash in search results" checkbox in the general preferences.  To solve this issue pretty simply, I check if this is a user initiated search (i.e. the search tab), in which case I don't omit anything. So for folders like Inbox trashed messages are still omitted. They are also still omitted for saved searches and tags - I discussed this with Josh and he thinks that difference makes sense.

One additional bug I encountered when fixing the trash issue is that in getFoldersToOmit it was getting the current search from appCtxt - that was not good, as when the search item list was generated, when coming from the search toolbar, the current search could have still been Inbox or anything else. Luckily we have a _currentSearch field in ZmMailListController, which is what I use instead and that is more accurate. It's not the first time I see a problem of using appCtxt.getCurrentSearch, expecting it to be something it's not.

http://bugzilla.zimbra.com/show_bug.cgi?id=93453

Copied from Perforce
 Change: 517655