Bug 93200
ShortCut Keys no longer work for web client when viewing message in a new window


Ha, well actually keyboard shortcuts do work, but not all of them. The 2 provided here do not work "P" and "M", but "R" (reply) and "F" (forward) for example work.

And again - this happens in any standalone message view (ZWC inner tab too, no need for new window).

It's a regression from bug 74060.

I added "M" and "P" (move and print), and also "T" (tag).

Whitelist was added in bug 74060 to avoid exceptions in some other shortcuts

I noticed that "M" (move) was still not working in the case of Child Window, and that is due to the past where we didn't allow move in child window. So I removed that restriction and it works. (bug 59162 is related).

http://bugzilla.zimbra.com/show_bug.cgi?id=93200

Copied from Perforce
 Change: 517972