Bug 93780 - Inline attachment is lost when format to switched to Plain text
It's a regression issue -- inline images were sent as regular attachments in 8.0 when compose mode was switched to plain text.
At revision #401883 (it addresses a number of bugs) in ZmComposeView.prototype._isInline(msg):

msg = msg || this._msg;
was changed to
msg = msg || this._origMsg;

In our case this._origMsg is undefined. Considering different approaches, I think as it's close to the release it's safest to pass this._msg to _isInline() function.

http://bugzilla.zimbra.com/show_bug.cgi?id=93780

Copied from Perforce
 Change: 518132