Bug 93596
The original email body is removed while replying


I wasn't able to get the body to be removed, but was able to get the "reply" button to not do anything, see comment #7 for repro steps.

I can only hope this fix would fix that other situation too.

The problem is due to clearing the list when changing folders, and clearing the list clears each item in the list, and ZmConv clear clears the conv from its messages.

This fix is to:
1. set the conv with an isInUse flag when it's shown in a standalone conv tab.
2. If conv.isInUse it's set, do not clear the conv when clearing the list.
2. Remove isInUse when the tab is removed.

3 was a bit tricky - had to add a so far not much used callback, ZmAppViewMgr.CB_POST_REMOVE, and implement _postRemoveCallback - this callback was added in bug 78891.

Also, testing it a little I discovered that there was a slight bug in bug 78891. I see the problem only when clicking "esc" after opening the tab - in that case ZmAppViewMgr.prototype.popView  is called without a viewId (which is fine), but then ZmAppViewMgr.CB_POST_REMOVE callback was taken from the "view" var instead of curView (I guess in other cases they are the same). view in the case of "esc" was this dummy "emptyView" as opposed to the current view (which is what is checked in the if anyway). I fixed that.

http://bugzilla.zimbra.com/show_bug.cgi?id=93596

Copied from Perforce
 Change: 518514