Bug 94166 - Creating new quick through 'add quick task' functionality while in different view(inProgress, Deferred) will display a duplicate task being created
Regression from bug 88323.

this.checkTaskReplenishListView() was added as a part of the fix for bug 88323 -- without it, when new task was created it wasn't getting selected.

Moved this.checkTaskReplenishView() inside the conditional block so that it's only called when  the item is a part of the current view.

One more issue: the item count was getting incremented on the top right (1-2 of 2, 1-3 of 3, etc) as quick tasks were created even though they were not getting added to the list. Added a check to see if the task is actually a part of the list before calling createNotify.

http://bugzilla.zimbra.com/show_bug.cgi?id=94166

Copied from Perforce
 Change: 519448