Bug 94406
Deleting message and clicking on undo link, message doesn't shows in Inbox,Required explicit refresh


NOTE: This happens only when using the checkbox to select (not 	selecting for view) or multi-select.

This is a regression from bug 74446. (unless it's a newer regression 	and I'm missing something out).

the problem is with checking the checkbox (without selecting the item for view). Same problem also with multi select so it's not solely a checkbox issue.

The problem is that the conv was not loaded in those cases, and so ZmConv.prototype._msgListChangeListener is not set as a change listener on the ZmMailList that contains the messages. Something like that, it's quite confusing. Because of that, ZmConvListView.prototype._changeListener was not called with a MOVE event in the CONV level (only the message level) when UNDOing.

BTW, that CONV MOVE event (when happens) is not handled in ZmConvListView.prototype._changeListener but in parent's _changeListener that's delegated at the end of ZmConvListView.prototype._changeListener if it was not handled.
Also, it re-runs the search in this case, that's how it adds the conv item, refreshing the entire list (to save on work using the	sort index to insert it).

My solution is to revert something that was done in bug 74446 (by 	the way, bug 85891 was also a result of this, mentioning it since my fix overrides that fix).
That is - bug 74446 replaced the CONV move in the logged actions (the ones that are used to undo later), with the individual MSG ones (since each one could be from a different folder), which sounds	good. But by not including the CONV move, there was no CONV MOVE
event as I mentioned above, thus the UI is not updated to display 	the conv again. The thing that confused me is that it does happen if the user clicked on the conv to view it, but that's through some other code path...

So what I do is include the CONV MOVE in addition to the MSG moves.
This seems to work. The CONV move results in the change listener called in a way that adds that line. And the MSG moves move the msgs to their respective folder.

http://bugzilla.zimbra.com/show_bug.cgi?id=94406

Copied from Perforce
 Change: 519767