Bug 94522
attachment disappears on adding signature while replying


See bug 82001. It dealt with same issue that happened when switching compose mode. So I piggyback on this solution.

However, made some change to that solution - changed back to having modeSwitch a local var passed as a param rather than an instance var as was done in bug 82001. This is clearer to read and debug and might prevent subtle race condition bugs.

And renamed it to keepAttachments in the params, as in the case of signatures modeSwitch does not apply, and also since it's more directly the purpose of this param.

http://bugzilla.zimbra.com/show_bug.cgi?id=94522

Copied from Perforce
 Change: 519869