Bug 90566
Selecting a conversation triggers mailbox server DosFilter


Different use-case and problem than the first one I fixed for this bug.

use-case is:
1. have a conv that's unread with N messages.
2. expand it.
3. click on it to select it for view in the reading pane.

result - a GetMsgRequest is fired for every msg.

Reason - in step 2 the SearchConvRequest is fired with needExp FALSE. (since we don't yet need the messages bodies, just the fragments, and some other meta data).
AFTER step 2, the conv is cached and marked as _loaded. But its messages are not loaded and no cached.

In step 3, since the conv is cached and loaded, it was not issuing a new SearchConvRequest. But since all the messages are unread, it then issues a GetMsgRequest for each.

Note that if you only do steps 1 and 3 (skip 2), it works fine, since then in step 3, it issues a SearchConvRequest with needExp TRUE, and so all the messages are fully loaded and cached and so no GetMsgRequests are issued (I do see a few, perhaps related to deleted ones on the conv I look at.... different story).

My solution is to keep a status _expanded for the conv, and if it's NOT expanded also force the search/load. This is similar to bug 75947. Specifically see comment:
https://bugzilla.zimbra.com/show_bug.cgi?id=75947#c1
But I do it locally in ZmConv, not involving ZmConvListController by re-using forceLoad, which was an option I thought about, but I think a change in just this class here is slightly cleaner?

http://bugzilla.zimbra.com/show_bug.cgi?id=90566

Copied from Perforce
 Change: 520105