Bug 94765
In-Reply-to sometimes does not contain the parent message-id, which causes broken thread on some mail client


This is related to bug 84453, and bug 82943. Both fixed some related issues.

The problem is the way we try to detect whether we need to set the IRT is not working so well. It's confusing code, msg.isDraft is not true even when saving a draft (the first time, probably timing of when it's set), but that is not really the issue either. msg.isReplied is true, correctly, so the line:
msg.irtMessageId = this._msg.messageId;
was executed.
There might be a solution that involves changing the condition so this line is not executed.

But what I did, and I think makes sense, is instead of using this._msg.messageId, I use this._origMsg.messageId. this._origMsg is the message you reply to, which is the only message that matters, I believe. (while this._msg is the saved draft message after you save a draft, which is not what we want).   I kept the condition as is since it's so confusing and fragile.

Also, if you edit a draft, the _origMsg is now the draft itself... weird, but not sure if that's changeable easily.
So updated the setting of irtMessageId to the _origMsg irtMessageId in that case, as opposed to the _origMsg.messageId.

http://bugzilla.zimbra.com/show_bug.cgi?id=94765

Copied from Perforce
 Change: 520673