Bug 93040
Multiple attachment of messages doesn't shows in Forward compose window


See comment #3 in the bug. The problem is we attempt to show bubbles for the forwarded messages (that are part of all the conversations we forward), in a way that's done before a draft is saved. The cause of the bug is that some of the messages (the ones we haven't viewed) are not cached since they are not loaded, and we try to get them from the cache.

This seems the simplest solution - save a draft right away when setting a view that gets messages to attach (params.msgIds). After the draft is saved, all the attached messages, including ones that are not cached, appear correctly as attachment bubbles.

http://bugzilla.zimbra.com/show_bug.cgi?id=93040

Copied from Perforce
 Change: 520942