Bug 95002
Clean up temporary msg attachment bubbles code when forwarding multiple messages


After my fix to bug 93040, I believe it's OK to get rid of the awkward design/code we had for showing the user that forwarded messages appear in the attachment bar.
We used to show them as bubbles, with all this interim and duplicate code and styling. But now since bug 93040, we save a draft right away, resulting in the Msg attachments appearing as regular attachment bubbles. So no need for all that code.

I made sure the case of "include original as attachment", that was also relying on that fake bubble, also saves a draft right away, so that works well too.

There's still some more that can be cleaned, and I started doing it (mainly in ZmComposeController, it does not really need _msgIds to be state var), but decided to keep it at this for now.

http://bugzilla.zimbra.com/show_bug.cgi?id=95002

Copied from Perforce
 Change: 521199