Bug: 91436 - User can't accept appointment which has an attachment in offline mode.
Same underlying problem as 94338, but I didn't get the fix for that quite right.  This problem comes from not having the cached appt associated with an invite.  That in turn comes from a problem with the delta sync updates, where the fix has the add/delete parameters swapped - sort of OK if it already existed and was cached (94338) but not so good if we are only adding a new entry.

http://bugzilla.zimbra.com/show_bug.cgi?id=91436

Copied from Perforce
 Change: 521268