Bug 80676
Send As Distribution Lists missing display name


Made the client side fix. This is btw for Send As and also Send On Behalf Of.

Also, originally the "Send on behalf of" internationalization was done like so:
displayName + " " + ZmMsg.sendOnBehalfOf + " "  + addr

This, I believe, is a no-no in i18n... (as some languages, the order might have to be reversed, so the whole phrase should be internationalized, not just the "on behalf of" words. So I fixed that while adding parts for the display name (if available).

http://bugzilla.zimbra.com/show_bug.cgi?id=80676

Copied from Perforce
 Change: 521415