Bug 94548
Persistent folder tree view can be accidentally modified by move shortcut / reuse of folder tree object


I actually tried to fix this issue all the way back in bug 72590, but it was not working for all cases.

My fix was to not persist the setting implicitly from the dialog overview. It's good, but not enough - we still set the setting in the memory settings. Note that the setting is per folder, not per [folder X overview], so even though those are 2 different overviews, they do as Phil mentioned, share the same collapse/expand setting.

So my fix to bug 72590, worked as long as no implicit save was done after the user used the dialog. But in our case it breaks in step 8 - that's where we notice that folder B changed from expanded to collapsed, and save the setting, including the one of folder a. (it's one setting variable for all, implementation detail).

BTW, it's not happening in step 7, due to the same bug - we set folder A's status in memory to expanded, when using the move dialog, so changing it to expanded in the tree makes it think nothing changed as far as setting, so no implicit save is done.

Long story short - my fix is to not even save the setting in memory, for dialog overview. NOTE that if the user opens the "move" dialog, collapse/expand some folders, closes it, and opens again - their status would remain, since we re-use the overview. So that's expected and good. It just would not affect the outside world (the mail tab overview), which is totally fine.

http://bugzilla.zimbra.com/show_bug.cgi?id=94548

Copied from Perforce
 Change: 521468