Bug 95064
Modify Ajax Client to pass CSRF token for REST API


Set the csrfToken as a hidden input field for the form used to upload a file for import. See also bug 92561 for the server-side part. Server-side still needs to be updated to support the post param that is a result of this hidden input field. (headers are impossible to pass in this case).

Note, I generate is lazily dynamically, mostly because the code that generates that form initially (ZmImportView) is a bit hard to understand.

http://bugzilla.zimbra.com/show_bug.cgi?id=95064

Copied from Perforce
 Change: 521471