Bug: 92592 - Offline - Drag and drop on mini calendar
Disable MiniCal as a drag and drop target when offline.  Note - drag and drop of a contact onto the minical will still highlight with green (valid), even though it won't do anything.  Suppressing that feedback is probably more work than its worth, since it would require adding an extra param for a drop target ('invalid when offline' or some such) and handling it in the drag and drop framework.

http://bugzilla.zimbra.com/show_bug.cgi?id=92592

Copied from Perforce
 Change: 521766