Bug 95258 - Weekly Appointment created using QuickAdd Appointment doesnot show specific recurrence pattern for particular day
Issue ::- In the 'CreateAppointment' request, that goes to the server, the recurrence rule is a jsonObject with values as {rule: {interval:{ival:1}, freq: WEE} }, it should be rather like {rule: {interval:{ival:1}, freq: WEE}, byday:{wkday:[{day:WE}]}} . The 'byday' property is missing in the jsonObject of the recurrence rule that causes this issue .

Details :- Whenever we are creating a recurring appointment,we are assigning value '1' to the repeatCustom attribute of the instance of ZmRecurrence. In this specific scenario, the attribute 'repeatCustom' to the ZmRecurrence's instance has default value 0 , and hence the jsonObject that we are constructing while saving the appointment, doesnot contain any attribute 'byday'. At code level, in function ZmRecurrence.prototype.setJson, presently we are making a check if value of repeatcustom attribute is 1 and appointment is weekly recurring one, add attribute 'byday'  to the recurrence rule. Refer line no. 124 (if (this.repeatCustom != "1") ) and line no. 146 else if (this.repeatType == ZmRecurrence.WEEKLY) {

Fix : - In the QuickAdd Appointment window, when the user makes an appointment, a weekly recurring one by selecting the value 'weekly' from the repeat dropdown, we assign a value '1' to the repeatCustom attribute of the instance of ZmRecurrence .

http://bugzilla.zimbra.com/show_bug.cgi?id=95258

Copied from Perforce
 Change: 522000