Bug 94488 - Email contact, right-click, go to URL option is missing
Changeset:

1. ZmOperation.js
2. ZmBaseController.js
3. ZmMsg.properties

Description:

This functionality got removed in bug 60017 (changelist: 499116).

Added the textKey to provide text while registering the Go To Url option in ZmOperation.
The Go To Url option is added to the bubble Context menu.
The given address is processed in the _goToUrlListener and opens the extracted domain in a new window.

http://bugzilla.zimbra.com/show_bug.cgi?id=94488

Copied from Perforce
 Change: 522206