Bug 24436
shared mailbox - save sent messages to the shared sent folder


Implement the UI preference for zimbraPrefDelegatedSendSaveTarget.

After discussing with Josh, Conrad and Jeff decided to move this to the "delegates" area under accounts.

Found some way to simply keep the definition as it was, but just move it to the primary account template, and by adding it to the prefs of the ACCOUNTS section, to accompany the vague "ZmSetting.ACCOUNTS" pref, it somehow works.

I also made the layout as Josh suggested, adding a td instead of the colspan=2 that was there (which also does for all browsers what bug 75482 did for IE8 and down, just a BTW... shouldn't cause a problem).

http://bugzilla.zimbra.com/show_bug.cgi?id=24436

Copied from Perforce
 Change: 522415