Bug: 93228 - "Checkout" a file from briefcase updates the timestamp of the version
The briefcase UI was using the modifiedDate (data.md) for display, which shows the last time the content or the metadata was modified.  So when a version is checked out, its modifiedDate is altered too.
I've changed the UI to use the data.d field, which is the last time the content was changed.

http://bugzilla.zimbra.com/show_bug.cgi?id=93228

Copied from Perforce
 Change: 522986