Bug 84371
Better organization of shared folders


Place all links (shared folders and entire shared mailboxes) AFTER the rest of the folders.

Josh noted that it was already doing this in Contacts, which was true (same was in calendar, tasks, and briefcase). Made me realize that there's all that duplicated code for the sortCompare method in Contacts, Calendar, Tasks and Briefcase.  The only difference between all those was that Contacts also put the "trash" folder AFTER all personal (local, non links) folders... which makes sense. Without it (as was in calendar, tasks and briefcase), it was sorter alphabetically and could end up with Trash as first folder, which is kinda awkward.

The main fix, of making sure the links in mail folder appear after the local folders, was as simple as an added "if" in ZmFolder.sortCompare.

But due to the above, I decided to refactor to reduce code duplication, and also to make sure Trash is at the end also on the places it wasn't (except of course Mail in which we have many system folders with a SORT_ORDER in which Trash is at the end of the system folders, but before the normal folders). I discussed this also with Josh and he agrees that my solution is correct.

BTW at first I tried to just make a change regarding the non mail Trash issue in ZmFolder.sortCompare, but that did not work since it was impossible to make it a correct sort (forget the 3 or so rules, but the one where if a < b and b < c then a < c was not followed). This was because we are doing something slightly weird, and that is we sort ALL the folders (mixed types) and then filter what we need to show in the view. Instead of first filtering, and then sorting. But I definitely didn't want to change this.  So to solve it I added a simple ZmFolder.sortCompareNonMail method to be used for non mail views, that passes a param indicating it to ZmFolder.sortCompare.

http://bugzilla.zimbra.com/show_bug.cgi?id=84371

Copied from Perforce
 Change: 523028