Bug 94867
Expand or '+'icon doesn't shows for user created DL


Taking Jeff's suggestion that "exp" is implicit in most places (particularly GetDistributionListsRequest, GetContactsRequest, GetDistributionList), I propose this simple fix of defaulting "exp" (or canExpand in the JS code) to true if it's not passed explicitly.
This solves the issue in the description, as well as the issue in bug 93902. Possibly other use-cases.

http://bugzilla.zimbra.com/show_bug.cgi?id=94867

Copied from Perforce
 Change: 523124