Integrate change 523533 from main to JUDASPRIEST
-------------
Bug 95064
Use XmlHttpRequest instead of form.submit(). This is in order to satisfy the server requiring an HTTP header named X-Zimbra-Csrf-Token to be the csrfToken.
Previously we were unable to do this, since we used form.submit() and that does not allow us to set any headers. So the previous solution was to pass it as a hidden form param, together with the File input that is uploaded. However Vishal said this is a problem for the server so here I'm using XmlHttpRequest to send the file to the server REST URL.
I use the ZmUploadManager for some code that's already there, but added some overriding params to it, since this is another case, there are already 2 different cases the Upload Manager supports, mail attachments, and briefcase uploads, and it branches on those 2 in the code. I simplified a little by adding a "url" param so I can pass the REST URL from the import, and a  stateChangeCallback param to be called from _handleUploadResponse so we can do our special thing in the import case.
I tried not to pass the callback param with the REST URL at first, since this works well with an iframe target for the form (but with XmlHttpRequest there's no target), as it returns full HTML that includes a script that calls the callback. But that, while working when no errors, was not returning the errors when there were errors (or warnings). So I eventually kept our code but set the iframe HTML from the response text as if that text was targeted to the IFRAME. This is the safest I think as it will still use the old tested code.

http://bugzilla.zimbra.com/show_bug.cgi?id=95064

Copied from Perforce
 Change: 523534