Bug 95448
Not able to see full window while changing Zoom levels in IE (Webclient)


This problem is only with IE11 in quirks mode, which is the case of IM (as we do not use standards mode there). But I don't think it hurts to have this code in JP as well just for easier reference, and consistency with IM for faster maintenance. It would do nothing on JP.

Added an AjxEnv.isIE11Quirks that detects whether the user is using IE11 in quirks mode (as is the case with IE11 on IM).

If this is the case, set a timer interval of 2 seconds to keep testing if the user zoomed in or out. (by testing a change in the value of window.screen.deviceXDPI which changes on zoom).  Interval of 2 seconds is a little arbitrary, but I believe that an average of 1 seconds to start resizing after zoom, or max 2 seconds, is more than satisfactory for the event of user changing the zoom. And performance-wise it should be completely insignificant.

http://bugzilla.zimbra.com/show_bug.cgi?id=95448

Copied from Perforce
 Change: 523576