Bug 95022
inline images missing during reply/forward compose


Relying on the ZmHtmlEditor addOnContentInitializedListener mechanism to do stuff in callbacks is not reliable, as ZmHtmlEditor.prototype.onLoadContent sometimes is called before those listeners are added (perhaps it's fixable), and possibly it might not be called at all (not sure at all).

Before bug 34874 it was just doing a timeout the first time so the editor has time to initialize. Bug 34874 replaced it with the above mentioned mechanism, but it either broke since then or was never reliable.

My fix is to have a timeout (for 3 seconds, a bit arbitrary, not too long but not too short) in addition to the listener, that does the same, so in case the onLoadContent thing didn't work, it will still do the 2 things that are done via listeners (_fixMultipartRelatedImages_onTimer, which is the reason for this bug, and also _saveComponentContent, which was probably also problematic when not called).

In addition, fixed _firstTimeFixImages to be set according to its name. Instead of undefined (false) at first, it should be "true" at first and "false" after the first time.

http://bugzilla.zimbra.com/show_bug.cgi?id=95022

Copied from Perforce
 Change: 523776